+++ core/core.php
@@ -1,22 +1,27 @@
 <?php
 
-namespace Kirby\Plugins\distantnative\oEmbed;
+namespace Kirby\distantnative\oEmbed;
 
 use A;
 use C;
 
+use Kirby\distantnative\Cache;
+use Kirby\distantnative\Thumb;
+
 class Core {
+
+  public $provider = null;
 
   public function __construct($url, $args = []) {
     $this->input = $url;
-    $this->cache = new Cache($url);
+    $this->cache = c::get('plugin.oembed.caching', true) ? new Cache('oembed', $url) : false;
 
     $this->load();
 
     if($this->data !== false) {
+      $this->options  = $this->options($args);
       $this->provider = $this->provider();
       $this->url      = new Url($this->data()->code);
-      $this->options  = $this->options($args);
     }
   }
 
@@ -25,19 +30,18 @@ class Core {
   // ================================================
 
   protected function load() {
-
     // get data from cache
-    if($this->cache->exists() && c::get('plugin.oembed.caching', true)) {
-      $this->data = $this->cache->get();
+    if($this->cache && $this->cache->exists()) {
+      return $this->data = $this->cache->get();
 
     // load data from source
     } else {
       $this->data = Data::get($this->input);
+    }
 
-      // cache the data
-      if($this->data && c::get('plugin.oembed.caching', true)) {
-        $this->cache->set($this->data, c::get('plugin.oembed.caching.duration', 24) * 60);
-      }
+    // cache the data
+    if($this->cache && $this->data) {
+      $this->cache->set($this->data, c::get('plugin.oembed.caching.duration', 24) * 60);
     }
   }
 
@@ -46,15 +50,13 @@ class Core {
   // ================================================
 
   protected function options($options) {
-    $defaults = [
+    return a::merge([
       'class'     => null,
       'thumb'     => null,
       'autoplay'  => c::get('plugin.oembed.video.autoplay', false),
       'lazyvideo' => c::get('plugin.oembed.video.lazyload', true),
       'jsapi'     => c::get('plugin.oembed.providers.jsapi', false),
-    ];
-
-    return a::merge($defaults, $options);
+    ], $options);
   }
 
 
@@ -64,9 +66,9 @@ class Core {
 
   public function thumb() {
     // if custom thumbnail is set
-    $thumb = $this->options['thumb'] ? $this->options['thumb'] : $this->image();
+    $thumb = $this->options['thumb'] ?: $this->image();
 
-    return new Thumb($thumb);
+    return $this->cache ? new Thumb('oembed', $thumb,       (c::get('plugin.oembed.caching.duration', 24) * 60 * 60)) : $thumb;
   }
 
 
@@ -78,7 +80,7 @@ class Core {
     $namespace = 'Kirby\Plugins\distantnative\oEmbed\Providers\\';
     $class     =  $namespace . $this->data()->providerName;
     $class     =  class_exists($class) ? $class : $namespace . 'Provider';
-    return new $class($this, $this->input);
+    return $this->provider ?: new $class($this, $this->input);
   }
 
 
@@ -99,10 +101,6 @@ class Core {
   }
 
   public function __toString() {
-    if($this->data === false) {
-      return Html::error($this->input);
-    }
-
-    return (string)new Html($this);
+    return !$this->data ? Html::error($this->input) : (string)new Html($this);
   }
 }