+++ core/providers/slideshare.php
@@ -2,18 +2,14 @@
 
 namespace Kirby\Plugins\distantnative\oEmbed\Providers;
 
-class Spotify extends Provider {
+class Slideshare extends Provider {
 
   protected function init() {
-    $this->getString('theme');
-    $this->getString('view');
     $this->getNumber('width');
     $this->getNumber('height');
   }
 
   public function code($code) {
-    $this->set('theme');
-    $this->set('view');
     $code = $this->setSizes($code);
     return $code;
   }
@@ -23,10 +19,8 @@ class Spotify extends Provider {
   //  Parameters for Panel Field Cheatsheet
   // ================================================
 
-  public function providerParameters() {
+  public function urlParameters() {
     return [
-      ['view', 'Set the view style (list/coverart)'],
-      ['theme', 'Set the theme (white/black)'],
       ['width', 'Set the width of the embed (e.g. 600)'],
       ['height', 'Set the height of the embed (e.g. 80)'],
     ];