+++ oembed.php
@@ -1,112 +1,35 @@
 <?php
 
-namespace Kirby\Plugins\distantnative\oEmbed {
+namespace Kirby\distantnative\oEmbed {
 
-require_once('lib/autoload.php');
+  use Kirby\distantnative\Autoloader;
+  require_once('core/lib/autoloader.php');
 
-$kirby    = kirby();
-$language = $kirby->site()->language();
-$language = $language ? $language->code() : null;
+  $kirby    = kirby();
+  $language = $kirby->site()->language();
 
-// ================================================
-//  Load components
-// ================================================
+  Autoloader::load([
+    'vendor'         => ['Embed/src/autoloader'],
+    'translations'   => ['en', $language ? $language->code() : null],
+    'core'           => ['core', 'data', 'url', 'html'],
+    'core/lib'       => ['cache', 'thumb'],
+    'core/providers' => ['provider', true],
+  ]);
 
-Autoloader::load([
-  'vendor'       => ['Embed/src/autoloader'],
-  'core'         => ['core', 'url', 'html'],
-  'lib'          => ['data', 'cache', 'thumb'],
-  'translations' => ['en', $language],
-  'providers'    => ['provider', true]
-]);
-
-
-
-// ================================================
-//  $page->video()->oembed()
-// ================================================
-
-$kirby->set('field::method', 'oembed', function($field, $args = []) {
-  return oembed($field->value, $args);
-});
-
-
-// ================================================
-//  (oembed: …)
-// ================================================
-
-$options = [
-  'class'     => 'string',
-  'thumb'     => 'string',
-  'autoload'  => 'bool',
-  'lazyvideo' => 'bool',
-  'jsapi'     => 'bool',
-];
-
-$kirby->set('tag', 'oembed', [
-  'attr' => array_keys($options),
-  'html' => function($tag) use($options) {
-    $args = [];
-
-    foreach($options as $option => $mode) {
-      if($mode === 'bool') {
-        if($tag->attr($option) === 'true')  $args[$option] = true;
-        if($tag->attr($option) === 'false') $args[$option] = false;
-      } elseif ($mode === 'string') {
-        if($tag->attr($option)) $args[$option] = $tag->attr($option);
-      }
-    }
-
-    return oembed($tag->attr('oembed'), $args);
-  }
-]);
-
-
-// ================================================
-//  Register panel field
-// ================================================
-
-$kirby->set('field', 'oembed', __DIR__ . DS . 'field');
-$kirby->set('route', [
-  'pattern' => 'api/plugin/oembed/preview',
-  'action'  => function() {
-    $oembed = oembed(get('url'), [
-      'lazyvideo' => true
-    ]);
-
-    $response = [];
-
-    if($oembed->data === false) {
-      $response['success'] = 'false';
-
-    } else {
-      $response['success']      = 'true';
-      $response['title']        = Html::removeEmojis($oembed->title());
-      $response['authorName']   = $oembed->authorName();
-      $response['authorUrl']    = $oembed->authorUrl();
-      $response['providerName'] = $oembed->providerName();
-      $response['providerUrl']  = $oembed->url();
-      $response['type']         = ucfirst($oembed->type());
-      $response['parameters']   = Html::cheatsheet($oembed->providerParameters());
-    }
-
-    if(get('code') === 'true') {
-      $response['code'] = (string)$oembed;
-    }
-
-    return \response::json($response);
-  },
-  'method'  => 'POST'
-]);
+  include('registry/field-method.php');
+  include('registry/tag.php');
+  include('registry/field.php');
+  include('registry/route.php');
 
 }
 
-namespace {
-  // ================================================
-  //  Global helper
-  // ================================================
 
+// ================================================
+//  Global helper
+// ================================================
+
+namespace {
   function oembed($url, $args = []) {
-    return new Kirby\Plugins\distantnative\oEmbed\Core($url, $args);
+    return new Kirby\distantnative\oEmbed\Core($url, $args);
   }
 }