+++ snippets/thumb.php
@@ -1,1 +1,0 @@
-<div class="kirby-plugin-oembed__thumb" style="background-image:url('<?= $url ?>');" title="<?= $alt ?>"><img src="<?= url(c::get('plugin.oembed.video.lazyload.btn', 'assets/plugins/oembed/images/play.png')) ?>" alt="" width="175" height="110" /></div>