+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [2.4.0](https://github.com/distantnative/oembed/releases/tag/2.4.0) (2016-09-04)
+- Improved: LazyVideo support and display for certain providers
+- Improved: Smarter sizing of LazyVideo play overlay
+- Improved: Preview label hidden in panel field
+- Improved: Panel field cheatsheet hidden by default, better cursor on hover
+- Improved: Removed duplicate assets for panel field
+- Improved: Updated vendor lib Embed to version 2.7
+- Improved: Internal file structure
+- Improved: Clearer namespacing, separation between plugin and lib
+- Fixed: Thumb caching for thumbs with url parameter
+- Fixed: Correct permissions for thumbs directories
+
 ## [2.3.2](https://github.com/distantnative/oembed/releases/tag/2.3.2) (2016-09-03)
 - Fixed: YouTube timecode URL parameter 't' did not work